exceed seat check on internal service
